requirements.txt
python-bitcoinrpc==1.0
python-dotenv==0.16.0
